requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==20.3.0
autopep8==1.5.4
cloudinary==1.26.0
configparser==5.0.2
coreapi==2.3.3
cryptography==3.4.7
decorator==5.0.7
Django==3.1
django-common-helpers==0.9.2
django-cors-headers==3.4.0
django-crontab==0.7.1
django-editorjs==0.2.1
django-extensions==3.1.3
django-heroku==0.3.1
django-phonenumber-field==5.2.0
django-templated-mail==1.1.1
django-tinymce==3.3.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
gunicorn==20.1.0
oauthlib==3.1.0
phonenumbers==8.12.27
pickleshare==0.7.5
Pillow==8.1.0
pkcs7==0.1.2
protobuf==3.17.3
psycopg2==2.8.5
psycopg2-binary==2.8.6
pycparser==2.20
pydantic==1.7.3
PyJWT==1.7.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.18.0
python-multipart==0.0.5
pytz==2020.1
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
threadpoolctl==2.1.0
unidecode==1.3.2
websocket-client==0.59.0
Werkzeug==1.0.1
whitenoise==5.2.0
zope.event==4.5.0
zope.interface==5.4.0
